# v16-e2e/front-end/docs/backend-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Customer Service Agent API",
        "description": "## AI-Powered Customer Service REST API\n\nThis API provides an intelligent customer service agent with:\n\n### 🤖 Core Features\n- **Conversational AI**: Natural language understanding powered by Azure OpenAI\n- **Multi-Turn Conversations**: Session-based dialogue management\n- **Order Management**: Query, track, and modify customer orders\n- **Email Communication**: Automated customer notifications\n- **Complaint Tracking**: MCP integration for complaint management\n\n### 🔧 Technical Capabilities\n- **Multi-Tenant**: Tenant isolation for session management\n- **Rate Limiting**: Configurable per-IP request throttling\n- **Request Tracing**: UUID-based request correlation\n- **Health Monitoring**: Liveness and readiness probes\n\n### 📚 Documentation\n- **Interactive Docs**: Available at `/docs` (Swagger UI)\n- **Alternative Docs**: Available at `/redoc` (ReDoc)\n- **OpenAPI Spec**: Available at `/openapi.json`\n\n### 🚀 Quick Start\n1. Check service health at `/health`\n2. Create a session at `/api/v1/agent/sessions`\n3. Send messages at `/api/v1/agent/messages`\n4. View session history at `/api/v1/sessions/{id}/history`\n",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
...
    },
    "paths": {
...
        "/agent/sessions": {
            "post": {
                "tags": [
                    "Agent"
                ],
                "summary": "Create Session",
                "description": "Create a new conversation session with the agent",
                "operationId": "create_session_agent_sessions_post",
                "parameters": [
                    {
                        "name": "x-request-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Request-Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSessionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateSessionResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Agent"
                ],
                "summary": "List Sessions",
                "description": "List all active conversation sessions",
                "operationId": "list_sessions_agent_sessions_get",
                "parameters": [
                    {
                        "name": "x-request-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Request-Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListSessionsResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/agent/messages": {
            "post": {
                "tags": [
                    "Agent"
                ],
                "summary": "Send Message",
                "description": "Send a message to the agent in a session",
                "operationId": "send_message_agent_messages_post",
                "parameters": [
                    {
                        "name": "x-request-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Request-Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendMessageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SendMessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/agent/sessions/{session_id}": {
            "delete": {
                "tags": [
                    "Agent"
                ],
                "summary": "Delete Session",
                "description": "Delete a conversation session",
                "operationId": "delete_session_agent_sessions__session_id__delete",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    },
                    {
                        "name": "x-request-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Request-Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteSessionResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/sessions/{session_id}": {
            "delete": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Delete Session",
                "description": "Delete a session.\n\n**Parameters:**\n- `session_id`: Session identifier\n- `tenant_id`: Tenant identifier (default: \"default\")\n\n**Returns:**\n- Confirmation message\n\n**Errors:**\n- `404`: Session not found",
                "operationId": "delete_session_api_v1_sessions__session_id__delete",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    },
                    {
                        "name": "tenant_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Tenant identifier",
                            "default": "default",
                            "title": "Tenant Id"
                        },
                        "description": "Tenant identifier"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/sessions": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "List Sessions",
                "description": "List all sessions for a tenant.\n\n**Parameters:**\n- `tenant_id`: Tenant identifier (default: \"default\")\n- `limit`: Maximum number of sessions (1-100, default: 50)\n- `offset`: Number of sessions to skip (default: 0)\n\n**Returns:**\n- List of session summaries",
                "operationId": "list_sessions_api_v1_sessions_get",
                "parameters": [
                    {
                        "name": "tenant_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Tenant identifier",
                            "default": "default",
                            "title": "Tenant Id"
                        },
                        "description": "Tenant identifier"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "description": "Maximum number of sessions to return",
                            "default": 50,
                            "title": "Limit"
                        },
                        "description": "Maximum number of sessions to return"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "description": "Number of sessions to skip",
                            "default": 0,
                            "title": "Offset"
                        },
                        "description": "Number of sessions to skip"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SessionListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateSessionRequest": {
                "properties": {
                    "session_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 100,
                                "minLength": 1
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Session Id",
                        "description": "Optional custom session ID (auto-generated if not provided)"
                    },
                    "tenant_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 100,
                                "minLength": 1
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tenant Id",
                        "description": "Optional tenant identifier for multi-tenancy"
                    },
                    "metadata": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Metadata",
                        "description": "Optional metadata to attach to the session"
                    }
                },
                "type": "object",
                "title": "CreateSessionRequest",
                "description": "Request model for creating a new conversation session."
            },
            "CreateSessionResponse": {
                "properties": {
                    "session_id": {
                        "type": "string",
                        "title": "Session Id",
                        "description": "Unique session identifier"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status",
                        "description": "Operation status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "Human-readable message"
                    },
                    "created_at": {
                        "type": "string",
                        "title": "Created At",
                        "description": "Session creation timestamp"
                    },
                    "metadata": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Metadata",
                        "description": "Session metadata"
                    }
                },
                "type": "object",
                "required": [
                    "session_id",
                    "status",
                    "message",
                    "created_at"
                ],
                "title": "CreateSessionResponse",
                "description": "Response model for session creation."
            },
            "DeleteSessionResponse": {
                "properties": {
                    "session_id": {
                        "type": "string",
                        "title": "Session Id",
                        "description": "Deleted session identifier"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status",
                        "description": "Operation status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "Human-readable message"
                    }
                },
                "type": "object",
                "required": [
                    "session_id",
                    "status",
                    "message"
                ],
                "title": "DeleteSessionResponse",
                "description": "Response model for session deletion."
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ListSessionsResponse": {
                "properties": {
                    "sessions": {
                        "items": {
                            "$ref": "#/components/schemas/SessionInfo"
                        },
                        "type": "array",
                        "title": "Sessions",
                        "description": "List of active sessions"
                    },
                    "total_count": {
                        "type": "integer",
                        "title": "Total Count",
                        "description": "Total number of sessions"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status",
                        "description": "Operation status"
                    }
                },
                "type": "object",
                "required": [
                    "sessions",
                    "total_count",
                    "status"
                ],
                "title": "ListSessionsResponse",
                "description": "Response model for listing sessions."
            },
...
            "SendMessageRequest": {
                "properties": {
                    "message": {
                        "type": "string",
                        "maxLength": 10000,
                        "minLength": 1,
                        "title": "Message",
                        "description": "User message to send to the agent"
                    },
                    "session_id": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 1,
                        "title": "Session Id",
                        "description": "Session ID for the conversation"
                    },
                    "tenant_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 100,
                                "minLength": 1
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tenant Id",
                        "description": "Optional tenant identifier"
                    },
                    "stream": {
                        "type": "boolean",
                        "title": "Stream",
                        "description": "Whether to stream the response (not yet implemented)",
                        "default": false
                    }
                },
                "type": "object",
                "required": [
                    "message",
                    "session_id"
                ],
                "title": "SendMessageRequest",
                "description": "Request model for sending a message to the agent."
            },
...
        }
    },
...
}